refactor: 테스트 domain 생성 메서드를 DomainUtils 로 분리
