+++ src/test/java/com/gugucon/shopping/order/domain/entity/OrderItemTest.java
@@ -1,15 +1,15 @@
 package com.gugucon.shopping.order.domain.entity;
 
-import com.gugucon.shopping.TestUtils;
-import com.gugucon.shopping.common.exception.ErrorCode;
-import com.gugucon.shopping.common.exception.ShoppingException;
-import com.gugucon.shopping.item.domain.entity.CartItem;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.junit.jupiter.api.Assertions.assertThrows;
+
+import com.gugucon.shopping.common.exception.ErrorCode;
+import com.gugucon.shopping.common.exception.ShoppingException;
+import com.gugucon.shopping.item.domain.entity.CartItem;
+import com.gugucon.shopping.utils.DomainUtils;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 
 @DisplayName("OrderItem 단위 테스트")
 class OrderItemTest {
@@ -18,7 +18,7 @@ class OrderItemTest {
     @DisplayName("OrderItem을 생성한다.")
     void create() {
         // given
-        final CartItem cartItem = TestUtils.createCartItem();
+        final CartItem cartItem = DomainUtils.createCartItem();
 
         // when & then
         assertThatNoException().isThrownBy(() -> OrderItem.from(cartItem));
@@ -29,7 +29,7 @@ class OrderItemTest {
     void createFail_soldOutProduct() {
         // given
         final CartItem cartItem = CartItem.builder()
-                .product(TestUtils.createSoldOutProduct("name", 1000))
+                .product(DomainUtils.createSoldOutProduct("name", 1000))
                 .quantity(1)
                 .memberId(1L)
                 .build();
@@ -44,7 +44,7 @@ class OrderItemTest {
     void createFail_tooManyQuantity() {
         // given
         final CartItem cartItem = CartItem.builder()
-                .product(TestUtils.createProduct(1))
+                .product(DomainUtils.createProduct(1))
                 .quantity(2)
                 .memberId(1L)
                 .build();
